[ARC-595] Add __main__ entrypoint to root library
branch: feature/ARC-595-mpyl-basic-cli
