+++ Pipfile
@@ -38,6 +38,7 @@ coverage = "*"
 python_version = "3.9"
 
 [scripts]
+cli = "python src/mpyl"
 build = "python setup.py install bdist bdist_wheel"
 create-graph = "pyreverse -o png src/mpyl"
 lint = "pylint --rcfile .pylintrc src/"
