+++ src/mpyl/cli/meta_info.py
@@ -34,12 +34,12 @@ VDB_LOGO = """
 
 
 class MetaInfo:
-    @staticmethod
-    def print_version():
-        print(f"Version {version_meta('mpyl')}")
+    @property
+    def version(self):
+        return f"Version {version_meta('mpyl')}"
 
-    @staticmethod
-    def print_about():
+    @property
+    def about(self):
         dist = distribution('mpyl')
         print(os.linesep.join(str(dist.metadata).split(os.linesep)[1:16]))
         print(VDB_LOGO)
